create get ERC721 tokens logic
