+++ components/03-organisms/TokensShelf.tsx
@@ -3,12 +3,10 @@ import { useAuthenticatedUser } from "@/lib/client/hooks/useAuthenticatedUser";
 import {
   getERC721TokensFromAddress,
   getERC20TokensFromAddress,
-  getERC721MetadataFromContractAddress,
 } from "@/lib/client/blockchain-utils";
 import { EthereumAddress, Token } from "@/lib/shared/types";
 import { TokensList } from "@/components/02-molecules";
 import { SelectUserIcon, SwapContext } from "@/components/01-atoms";
-import { usePonder } from "@/lib/client/hooks/usePonder";
 import { useContext, useEffect, useState } from "react";
 import { useTheme } from "next-themes";
 import { useNetwork } from "wagmi";
@@ -34,9 +32,7 @@ interface TokensShelfProps {
 export const TokensShelf = ({ address, variant }: TokensShelfProps) => {
   const { chain } = useNetwork();
   const [allTokensList, setAllTokensList] = useState<Token[]>([]);
-  const [allTokensListByPonder, setAllTokensListByPonder] = useState<Token[]>(
-    [],
-  );
+
   const [tokensQueryStatus, setTokensQueryStatus] = useState<TokensQueryStatus>(
     TokensQueryStatus.EMPTY_QUERY,
   );
@@ -49,8 +45,6 @@ export const TokensShelf = ({ address, variant }: TokensShelfProps) => {
     destinyChain,
     ponderFilterStatus,
   } = useContext(SwapContext);
-
-  const { erc721AskSwaps } = usePonder();
 
   const getUserTokens = async () => {
     const chainId = authenticatedUserAddress?.equals(address)
@@ -88,45 +82,9 @@ export const TokensShelf = ({ address, variant }: TokensShelfProps) => {
     }
   };
 
-  // Here we will transform to Token the ERC721 | ERC20
-  const getUserTokensByPonder = async () => {
-    const chainId = authenticatedUserAddress?.equals(address)
-      ? chain?.id
-      : ChainInfo[destinyChain].id;
-
-    let queriedTokens: Token[] = [];
-
-    if (address && chainId) {
-      console.log("erc721AskSwaps", erc721AskSwaps);
-      Promise.all([
-        getERC721MetadataFromContractAddress(chainId, erc721AskSwaps).then(
-          (tokens) => {
-            console.log("tokens s s =", tokens);
-            queriedTokens = [...queriedTokens, ...tokens];
-          },
-        ),
-        //  getERC20MetadataFromContractAddress(address, chainId).then(
-        //    (tokens) => {
-        //      queriedTokens = [...queriedTokens, ...tokens];
-        //      tokensCount = tokensCount + tokens.length;
-        //    },
-        //  ),
-      ])
-        .catch(() => {
-          queriedTokens = [];
-        })
-        .finally(() => {
-          if (chainId) {
-          } else {
-            setAllTokensListByPonder(queriedTokens);
-          }
-        });
-    }
-  };
-
   useEffect(() => {
     getUserTokens();
-    getUserTokensByPonder();
+    // getUserTokensByPonder();
   }, [address, chain, destinyChain, ponderFilterStatus]); // PonderFilterStatus & destinyChain addded probably must be moved to another place
 
   const conditionallyCleanTokensList = (condition: boolean) => {
